rough outline of userui to db

+++ src/main/java/dao/interfaces/IBaseDAO.java
@@ -15,7 +15,7 @@ public interface IBaseDAO<T> {
      */
     int create(T o) throws SQLException;
 
-    T getById(long id) throws SQLException;
+    T getById(int id) throws SQLException;
 
     List<T> getAll() throws SQLException;
